Remove goto's. Remove dead code.
Goto's are bad. This refactors the code to remove the use of goto's
within the code. One of the goto's found was in a large chunk of
code that had been commented out. Removed that entire section of code
since that is what source control is for. If that section of code was
an experimental branch, then put that code in a branch, not as
disabled code.